+++ CHANGELOG.md
@@ -1,9 +1,35 @@
 # Change Log
 
-All notable changes to the "torrent-helper" extension will be documented in this file.
+All notable changes to the "smallit-vue-helper" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- コンポーネント用スニペットを追加予定
+
+### Fixed
+
+- 一部のスニペットの prefix を修正予定
+
+## [0.0.2] - 2025-04-24
+
+### Added
+
+- 「to-xxxx」でもスニペットを呼び出せるようになりました。
+
+### Fixed
+
+- Vueファイル内の <template></template> や <script></script> などのタグ内で動作しないバグを修正しました。
+
+### Removed
+
+- 一部不要なファイルを削除しました。
+
+## [0.0.1] - 2025-04-21
+
+### Added
+
+- プレリリース（https://github.com/smallit-torrent/smallit-vue-helper/releases/tag/v0.0.1）